requirements.txt
streamlit
pandas
python-docx
unidecode
openpyxl
docxcompose
pypdf
docx2pdf
pyyaml
reportlab
jinja2
